Rework selection code
In this process, we lost "injection" of new totally random genomes
but it's unclear if that's something we should care about

We also somewhat awkwardly have combined the parameter that controls
truncation selection with elitism. It's unlikely that both of these
things could be well tuned with the same value.

But let's deal with that when we implement something better than
truncation selection